debug

+++ portiloop_detector/pareto_search.py
@@ -400,7 +400,7 @@ class SurrogateModel(nn.Module):
                   config_dict["dilation_conv"],
                   config_dict["dilation_pool"]]
 
-        x_tensor = torch.tensor(x_list).to(self.device)
+        x_tensor = torch.tensor(x_list, dtype=torch.long).to(self.device)
 
         x_tensor = F.relu(self.d1(self.fc1(x_tensor)))
         x_tensor = F.relu(self.d2(self.fc2(x_tensor)))
